Ignore page load failures initiated by Click action.
If the click finishes, it's finished.
The fact that the expected destination page didn't load it's irrelevant.

This implies that, in case a "onLoadFinished" occurred
but resulting "status === 'fail'" we have to ignore it.

This is how Chrome and Firefox Drivers seem to behave…

Fixes #202.
Conflicts:
	test/java/src/test/java/ghostdriver/ElementMethodsTest.java

(cherry-picked from d0615a547f5a036df3134ef946c33d972c384aac)